# _docs/swagger/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "back-end part for mobile application.",
        "title": "Intruct API",
        "contact": {
            "name": "Ruslan's Support",
            "url": "https://t.me/Ruslan20007",
            "email": "ruslanrbb8@gmail.com"
        },
        "version": "1.0"
    },
    "host": "intruct.com",
    "basePath": "/api",
    "paths": {
...
        "/course": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new course with title, description, file and language\nPrerequisites: User must be authenticated",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course"
                ],
                "summary": "Create course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Course file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Course language",
                        "name": "language",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/course.createCourseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation failed",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
        "/courses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of courses with optional filter for user's courses\nPrerequisites: User must be authenticated",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course"
                ],
                "summary": "List courses",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter user's courses (true/false)",
                        "name": "in_mine",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/course.listCoursesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
        "/courses/{course_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get course details including modules and lessons structure\nPrerequisites: User must be authenticated. Course must be published or user must be the owner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course"
                ],
                "summary": "Get course",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/course.getCourseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid course ID",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Authentication required or user not registered",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "403": {
                        "description": "Access forbidden - course not published",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
        "/lessons/{lesson_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get lesson details including content and quizzes with options\nPrerequisites: User must be authenticated. User must have access to the course and have reached this lesson in progression",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lesson"
                ],
                "summary": "Get lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID",
                        "name": "lesson_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lesson.getLessonResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid lesson ID",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "401": {
                        "description": "Authentication required or user not registered",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "403": {
                        "description": "Access forbidden - lesson locked or course access denied",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "404": {
                        "description": "Lesson not found",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.ErrorStruct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "course.courseItem": {
            "description": "courseItem represents a course in the list",
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                },
                "description": {
                    "type": "string",
                    "example": "Learn Go programming language"
                },
                "finished_lessons": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_in_mine": {
                    "type": "boolean",
                    "example": false
                },
                "is_mine": {
                    "type": "boolean",
                    "example": false
                },
                "is_public": {
                    "type": "boolean",
                    "example": true
                },
                "lessons_number": {
                    "type": "integer",
                    "example": 10
                },
                "title": {
                    "type": "string",
                    "example": "Introduction to Go"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                }
            }
        },
        "course.createCourseResponse": {
            "description": "createCourseResponse returns created course ID",
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "course.getCourseResponse": {
            "description": "getCourseResponse returns course details with modules and lessons",
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                },
                "description": {
                    "type": "string",
                    "example": "Learn Go programming language"
                },
                "finished_lessons": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_in_mine": {
                    "type": "boolean",
                    "example": false
                },
                "is_mine": {
                    "type": "boolean",
                    "example": false
                },
                "is_public": {
                    "type": "boolean",
                    "example": true
                },
                "lessons_number": {
                    "type": "integer",
                    "example": 10
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/course.moduleItem"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Introduction to Go"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                }
            }
        },
        "course.lessonItem": {
            "description": "lessonItem represents a lesson in the module",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                },
                "description": {
                    "type": "string",
                    "example": "Learn about basic data types"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "serial_number": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Variables and Types"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                }
            }
        },
        "course.listCoursesResponse": {
            "description": "listCoursesResponse returns list of courses",
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/course.courseItem"
                    }
                }
            }
        },
        "course.moduleItem": {
            "description": "moduleItem represents a module in the course",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                },
                "description": {
                    "type": "string",
                    "example": "Introduction to the basics"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/course.lessonItem"
                    }
                },
                "serial_number": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Getting Started"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                }
            }
        },
...
        "httputils.ErrorStruct": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "lesson.getLessonResponse": {
            "description": "getLessonResponse returns lesson details with quizzes and options",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "# Variables in Go\n\nIn Go, variables are..."
                },
                "course_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                },
                "description": {
                    "type": "string",
                    "example": "Learn about basic data types"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "module_id": {
                    "type": "integer",
                    "example": 1
                },
                "quizzes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lesson.quizItem"
                    }
                },
                "serial_number": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Variables and Types"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                }
            }
        },
        "lesson.quizItem": {
            "description": "quizItem represents a quiz in the lesson",
            "type": "object",
            "properties": {
                "correct_index": {
                    "type": "integer",
                    "example": 0
                },
                "created_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "0",
                        "null",
                        "undefined",
                        "nil"
                    ]
                },
                "question": {
                    "type": "string",
                    "example": "What is the zero value of an int?"
                },
                "serial_number": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-09-09T10:10:10Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}